Implement group separation for different forces
